requirements.txt
Django>=4.2
djangorestframework>=3.14
celery>=5.3
redis>=4.5
requests>=2.31
pytest>=7.0
pytest-django>=4.5
gunicorn>=20.1
django-cors-headers>=4.0
whitenoise>=6.0
psycopg2-binary>=2.9
drf-spectacular>=0.27.0